refactor: Move all BoundsDiscovery code into a separate file
